fix: moved DNSRecordManager
