+++ src/main/java/eu/divum/divumbackend/services/CloudflareDNSRecordManager.java
@@ -8,6 +8,7 @@ import eu.divum.divumbackend.exceptions.HTTPRequestException;
 
 import eu.divum.divumbackend.exceptions.cloudflare.CloudflareAPIException;
 
+import eu.divum.divumbackend.services.implementations.DNSRecordManager;
 import lombok.RequiredArgsConstructor;
 
 import org.apache.commons.validator.routines.InetAddressValidator;
