+++ src/main/java/eu/divum/divumbackend/services/implementations/DNSRecordManager.java
@@ -1,4 +1,4 @@
-package eu.divum.divumbackend.services;
+package eu.divum.divumbackend.services.implementations;
 
 public interface DNSRecordManager {
     String create(String domain, String ipAddress);
